requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.8.0
transformers>=4.30.0
torch>=2.0.0
# Optional: for better performance on VPS
psutil>=5.9.0
# Optional: for database if you expand later
# sqlite3 is included in Python standard library
